Move common test function into own create test_utils
