+++ core_api/Cargo.toml
@@ -7,3 +7,6 @@ authors = ["Eric Harris-Braun <eric@harris-braun.com>"]
 holochain_core = { path = "../core" }
 holochain_dna = { path = "../dna" }
 holochain_agent = { path = "../agent" }
+
+[dev-dependencies]
+test_utils = { path = "../test_utils"}
