+++ core_api/src/lib.rs
@@ -55,13 +55,12 @@ hc.stop().expect("couldn't stop the app");
 extern crate holochain_agent;
 extern crate holochain_core;
 extern crate holochain_dna;
+#[cfg(test)]
+extern crate test_utils;
 
 use holochain_core::{
-    context::Context,
-    error::HolochainError,
-    instance::Instance,
-    nucleus::{call_and_wait_for_result, Action::*, FunctionCall},
-    state::{Action::*, State},
+    context::Context, error::HolochainError, instance::Instance,
+    nucleus::{call_and_wait_for_result, Action::*, FunctionCall}, state::{Action::*, State},
 };
 use holochain_dna::Dna;
 use std::sync::Arc;
@@ -144,11 +143,9 @@ mod tests {
     use holochain_core::context::Context;
     use holochain_core::logger::Logger;
     use holochain_core::persister::SimplePersister;
-    use holochain_core::test_utils::{
-        create_test_dna_with_wasm, create_test_dna_with_wat, test_wasm_from_file,
-    };
     use std::fmt;
     use std::sync::{Arc, Mutex};
+    use test_utils::{create_test_dna_with_wasm, create_test_dna_with_wat, test_wasm_from_file};
 
     #[derive(Clone)]
     struct TestLogger {
